set cheap sourcemap for dev task
