+++ config/webapp.webpack.config.js
@@ -1,0 +1,332 @@
+const fs = require("fs");
+const path = require("path");
+const ProgressPlugin = require("webpack/lib/ProgressPlugin");
+const CircularDependencyPlugin = require("circular-dependency-plugin");
+const HtmlWebpackPlugin = require("html-webpack-plugin");
+
+const autoprefixer = require("autoprefixer");
+const postcssUrl = require("postcss-url");
+const cssnano = require("cssnano");
+
+const { NoEmitOnErrorsPlugin, SourceMapDevToolPlugin, NamedModulesPlugin } = require("webpack");
+const { GlobCopyWebpackPlugin, NamedLazyChunksWebpackPlugin, BaseHrefWebpackPlugin } = require("@angular/cli/plugins/webpack");
+const { CommonsChunkPlugin } = require("webpack").optimize;
+const { AotPlugin } = require("@ngtools/webpack");
+
+const nodeModules = path.join(process.cwd(), "node_modules");
+const realNodeModules = fs.realpathSync(nodeModules);
+const genDirNodeModules = path.join(process.cwd(), "src", "$$_gendir", "node_modules");
+const entryPoints = ["inline", "polyfills", "sw-register", "styles", "vendor", "main"];
+const minimizeCss = false;
+const baseHref = "";
+const deployUrl = "";
+const postcssPlugins = function () {
+  // safe settings based on: https://github.com/ben-eb/cssnano/issues/358#issuecomment-283696193
+  const importantCommentRe = /@preserve|@license|[@#]\s*source(?:Mapping)?URL|^!/i;
+  const minimizeOptions = {
+    autoprefixer: false,
+    safe: true,
+    mergeLonghand: false,
+    discardComments: { remove: (comment) => !importantCommentRe.test(comment) }
+  };
+  return [
+    postcssUrl({
+      url: (URL) => {
+        // Only convert root relative URLs, which CSS-Loader won"t process into require().
+        if (!URL.startsWith("/") || URL.startsWith("//")) {
+          return URL;
+        }
+        if (deployUrl.match(/:\/\//)) {
+          // If deployUrl contains a scheme, ignore baseHref use deployUrl as is.
+          return `${deployUrl.replace(/\/$/, "")}${URL}`;
+        }
+        else if (baseHref.match(/:\/\//)) {
+          // If baseHref contains a scheme, include it as is.
+          return baseHref.replace(/\/$/, "") +
+            `/${deployUrl}/${URL}`.replace(/\/\/+/g, "/");
+        }
+        else {
+          // Join together base-href, deploy-url and the original URL.
+          // Also dedupe multiple slashes into single ones.
+          return `/${baseHref}/${deployUrl}/${URL}`.replace(/\/\/+/g, "/");
+        }
+      }
+    }),
+    autoprefixer(),
+  ].concat(minimizeCss ? [cssnano(minimizeOptions)] : []);
+};
+
+module.exports = {
+  "resolve": {
+    "extensions": [
+      ".ts",
+      ".js",
+      ".css"
+    ],
+    "modules": [
+      "./node_modules",
+      "./node_modules"
+    ],
+    "symlinks": true
+  },
+  "resolveLoader": {
+    "modules": [
+      "./node_modules",
+      "./node_modules"
+    ]
+  },
+  "entry": {
+    "main": [
+      "./plugin/webapp/src/main.ts"
+    ],
+    "polyfills": [
+      "./plugin/webapp/src/polyfills.ts"
+    ],
+    "styles": [
+      "./plugin/webapp/src/styles.css"
+    ]
+  },
+  "output": {
+    "path": path.join(process.cwd(), "dist", "app"),
+    "filename": "[name].bundle.js",
+    "chunkFilename": "[id].chunk.js"
+  },
+  "module": {
+    "rules": [
+      {
+        "enforce": "pre",
+        "test": /\.js$/,
+        "loader": "source-map-loader",
+        "exclude": [
+          /(\\|\/)node_modules(\\|\/)/
+        ]
+      },
+      {
+        "test": /\.html$/,
+        "loader": "raw-loader"
+      },
+      {
+        "test": /\.(eot|svg|cur)$/,
+        "loader": "file-loader?name=[name].[hash:20].[ext]"
+      },
+      {
+        "test": /\.(jpg|png|webp|gif|otf|ttf|woff|woff2|ani)$/,
+        "loader": "url-loader?name=[name].[hash:20].[ext]&limit=10000"
+      },
+      {
+        "exclude": [
+          path.join(process.cwd(), "plugin/webapp/src/styles.css")
+        ],
+        "test": /\.css$/,
+        "use": [
+          "exports-loader?module.exports.toString()",
+          {
+            "loader": "css-loader",
+            "options": {
+              "sourceMap": false,
+              "importLoaders": 1
+            }
+          },
+          {
+            "loader": "postcss-loader",
+            "options": {
+              "ident": "postcss",
+              "plugins": postcssPlugins
+            }
+          }
+        ]
+      },
+      {
+        "exclude": [
+          path.join(process.cwd(), "plugin/webapp/src/styles.css")
+        ],
+        "test": /\.scss$|\.sass$/,
+        "use": [
+          "exports-loader?module.exports.toString()",
+          {
+            "loader": "css-loader",
+            "options": {
+              "sourceMap": false,
+              "importLoaders": 1
+            }
+          },
+          {
+            "loader": "postcss-loader",
+            "options": {
+              "ident": "postcss",
+              "plugins": postcssPlugins
+            }
+          },
+          {
+            "loader": "sass-loader",
+            "options": {
+              "sourceMap": false,
+              "precision": 8,
+              "includePaths": []
+            }
+          }
+        ]
+      },
+      {
+        "include": [
+          path.join(process.cwd(), "plugin/webapp/src/styles.css")
+        ],
+        "test": /\.css$/,
+        "use": [
+          "style-loader",
+          {
+            "loader": "css-loader",
+            "options": {
+              "sourceMap": false,
+              "importLoaders": 1
+            }
+          },
+          {
+            "loader": "postcss-loader",
+            "options": {
+              "ident": "postcss",
+              "plugins": postcssPlugins
+            }
+          }
+        ]
+      },
+      {
+        "include": [
+          path.join(process.cwd(), "plugin/webapp/src/styles.css")
+        ],
+        "test": /\.scss$|\.sass$/,
+        "use": [
+          "style-loader",
+          {
+            "loader": "css-loader",
+            "options": {
+              "sourceMap": false,
+              "importLoaders": 1
+            }
+          },
+          {
+            "loader": "postcss-loader",
+            "options": {
+              "ident": "postcss",
+              "plugins": postcssPlugins
+            }
+          },
+          {
+            "loader": "sass-loader",
+            "options": {
+              "sourceMap": false,
+              "precision": 8,
+              "includePaths": []
+            }
+          }
+        ]
+      },
+      {
+        "test": /\.ts$/,
+        "loader": "@ngtools/webpack",
+        "exclude": [/node_modules/, /\.spec.ts?$/]
+      }
+    ]
+  },
+  "plugins": [
+    new NoEmitOnErrorsPlugin(),
+    new GlobCopyWebpackPlugin({
+      "patterns": [
+        "assets",
+        "favicon.ico"
+      ],
+      "globOptions": {
+        "cwd": path.join(process.cwd(), "src"),
+        "dot": true,
+        "ignore": "**/.gitkeep"
+      }
+    }),
+    new ProgressPlugin(),
+    new CircularDependencyPlugin({
+      "exclude": /(\\|\/)node_modules(\\|\/)/,
+      "failOnError": false
+    }),
+    new NamedLazyChunksWebpackPlugin(),
+    new HtmlWebpackPlugin({
+      "template": "./plugin/webapp/src/index.html",
+      "filename": "index.html",
+      "hash": false,
+      "inject": true,
+      "compile": true,
+      "favicon": false,
+      "minify": false,
+      "cache": true,
+      "showErrors": true,
+      "chunks": "all",
+      "excludeChunks": [],
+      "title": "Webpack App",
+      "xhtml": true,
+      "chunksSortMode": function sort(left, right) {
+        let leftIndex = entryPoints.indexOf(left.names[0]);
+        let rightindex = entryPoints.indexOf(right.names[0]);
+        if (leftIndex > rightindex) {
+          return 1;
+        }
+        else if (leftIndex < rightindex) {
+          return -1;
+        }
+        else {
+          return 0;
+        }
+      }
+    }),
+    new BaseHrefWebpackPlugin({}),
+    new CommonsChunkPlugin({
+      "name": [
+        "inline"
+      ],
+      "minChunks": null
+    }),
+    new CommonsChunkPlugin({
+      "name": [
+        "vendor"
+      ],
+      "minChunks": (module) => {
+        return module.resource
+          && (module.resource.startsWith(nodeModules)
+            || module.resource.startsWith(genDirNodeModules)
+            || module.resource.startsWith(realNodeModules));
+      },
+      "chunks": [
+        "main"
+      ]
+    }),
+    new CommonsChunkPlugin({
+      "name": [
+        "main"
+      ],
+      "minChunks": 2,
+      "async": "common"
+    }),
+    new NamedModulesPlugin({}),
+    new AotPlugin({
+      "mainPath": "main.ts",
+      "replaceExport": false,
+      "hostReplacementPaths": {
+        "environments/environment.ts": "environments/environment.ts"
+      },
+      "exclude": [],
+      "tsConfigPath": "plugin/webapp/src/tsconfig.app.json",
+      "skipCodeGeneration": true
+    })
+  ],
+  "node": {
+    "fs": "empty",
+    "global": true,
+    "crypto": "empty",
+    "tls": "empty",
+    "net": "empty",
+    "process": true,
+    "module": false,
+    "clearImmediate": false,
+    "setImmediate": false
+  },
+  "devServer": {
+    "historyApiFallback": true
+  }
+};
