+++ package.json
@@ -4,7 +4,7 @@
   "description": "Wrap the needed tools required to cook the StravistiX Web Plugin",
   "scripts": {
     "build": "rimraf dist && webpack",
-    "dev": "rimraf dist && webpack -w",
+    "dev": "npm run build -- -w --env.dev",
     "test:core": "karma start ./karma.conf.js",
     "test:webapp": "karma start ./plugin/webapp/karma.conf.js"
   },
@@ -123,6 +123,6 @@
     "url-loader": "^0.5.9",
     "webpack": "^3.5.5",
     "webpack-dev-server": "^2.7.1",
-    "worker-loader": "^2.2.0"
+    "worker-loader": "^0.8.1"
   }
 }
